python-developer/07-flask/szkolenie-06-rest-api/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
requests
